Simplify the inactivity and sleeptimer
Remove the direct actions to go to deep standby/shutdown as this is not
really required anymore die to the standby timeout that allows the
receiver to go to deepstandby/shutdown later. This allows you when the box
goes by accident to deepstamdby that you quickly can undo it as there is
still some time added in standby. This also makes it much easier to avoid
going to deep standby when streaming is ongoing. I added also 1, 5 and 10
minutes to the standby to deepstandby timer.

From Commit [4e30af] Authored by: littlesat 2014-06-07