SoftwareManager: Add options to put plugin in setup menu
And also for the blue button menu. When it appears in the setup
menu and/or the blue button menu it will disappear from the plugin
menu, but for blue button and disappear from plugin menu a restart
of E2 is required. Default is on pluginmenu.

From Commit [7ea903] Authored by: Littlesat 2012-07-17