ChannelSelection: Only use one history buffer
As it was made before the history_pos was not changed when the history
buffer was swapped from radio to TV. This might lead to a gsod. I could
change the code to fix this but I thought it is even better to have just
one common history buffer.

From Commit [9062b6] Authored by: Littlesat 2014-07-02