Make config for default EPG independend from translations
By storing the path to the plugin in a config instead of plugin
description/name which was translatable. Use default names for standard EPG types when it is not within a plugin. A user needs to re-initiate their default plugin configs as the previous ones become invalid

From Commit [f3385f] Authored by: littlesat 2014-08-16