Correctly revert to current tv or radio bouquet
When e.g. we are in TV mode and the service is not available in any
bouquet then the TV mode bouquet was reverted and the radio bouquets were
focused to the latest radio bouquet. Now we always revert to the correct
tv or radio bouquet.

Fron Commit [0dccd6] Authored by: Littlesat 2014-08-30