ChoiceList: Only load PNG when key is defined
As suggested here:
http://openpli.org/forums/topic/28532-merge-requests-for-plis-git/page-32#entry441828

Use a format-string instead of concatenation. This allows any printable
object to be used as key name (e.g. numbers) instead of forcing key to be
a string, it's more efficient and it won't crash when key is something else.

From Commit [14c854] Authored by: Mike Looijmans 2014-09-07