Hotkey: When infobarfunction does not exist search for other key
When e.g. historyback was assigned to the left_cross key than in the
normal infobar it performs fine, but in the movie infobar nothing happens
as historyback does not exist. Now when an inforbar function does not
exist it reverts back to the usual function.

It could be considered to perform this for more Hotkey functions...

From Commit [f9cf4c] Authored by: littlesat 2014-09-26