ChannelSelection: Do not disable movemode when bouquet is changed
In addition corrigate the Activate/Disactivate move mode message in the
context menu. When we are in a list where we cannot replace entries you
select the entry instead of marking the service. Entries will only be
marked when we are in a userbouquet, alternatives overview list or the
userbouquet overview list.

From Commit [9007e6] Authored by: littlesat 2014-10-17