Satfinder: Hopefully final solution for predeftransponder list
As still from switching from different modes (S2/T/C) we still can get a
GSOD because changing the predef/manual config transponder was on the
retune notification list. It still does rebuild the config and because
after rebuilding the config also a retune is performed it can be removed
from the retune list avoiding GSOD when a type of transponder list is not
build yet.

Sorry for my previous two commits as they where not really required.

From Commit [761b27] Authored by: littlesat 2014-10-24