Language: Add GST_SUBTITLE_ENCODING failback environment variable
The GST_SUBTITLE_ENCODING environment variable will be checked for an encoding to use if input subtitles are not in UTF-8 encoding.
By default if that is not set either, ISO-8859-15 is assumed, that's why we are setting this as default failback.
Additionally gconv module for character set (e)glibc-gconv-iso8859-XX must be installed.

From Commit [ca521f] Authored by: Athanasios Oikonomou 2014-10-18
                     Committed by: Erik Slagter 2014-10-31