Revert "Advanced functions(edit/delete/disable) repeated timers in EPG list"
This reverts commit dab664bde057328ffd45be1f4b786a72acd0904e.

As it has unexpected side effects e.g. when selecting a overlap (half icon of repeated timer) you should be able to program a new timer and also some icons seems to be printed double (repeated and not repeated) when there is overlap.

From Commit [c1488f] Authored by: littlesat 2014-11-20