servicemp3: add progressive download buffering support for streams
-always enable progresive download buffering for streams,
whether or not a file buffer will actually be used depends
on the type and format of the stream
-always enable 'normal' buffering for streams as well
-ignore buffering messages, unless 'buffer=1' is specified
-put audio/video sinks in sync mode for streams.
This gives a more predictible dataflow, avoiding that
gstreamer buffer levels drop to 0 as soon as playback starts.

NOTE: we always use buffering, but by default we assume the
buffer levels will stay on a healthy level at all times.
So we assume we do not need to pause playback in order to fill
the buffer.
If the stream needs to be paused in order for the buffer
to reach (and keep) a healthy level, specify 'buffer=1'
in the stream url.

From Commit [e1e162] Authored by: pieterg 2012-08-18