Also hide buffering status when playing audio (radio) streams
Previously the detection if a stream was started was by detecting the VideoResolution
This of course does not work for audio streams. Therefore I added a evGstreamerPlayStarted
event to detect if a stream is actually starting to play which is now used to hide
the buffer indicator.

Authored by: Littlesat Littlesat 2014-11-30