dvb: use runtime DVB_API_VERSION check
we cannot rely on a compiletime check, because
the linux-libc-headers we're building against
might differ from the actual kernel headers.
So first do a compiletime check for the availability
of defines and enums (in linux-libc-headers),
followed by a runtime check for the actual kernel
support (where applicable).

From Commit [5f461d] Authored by: pieterg 2014-12-1