ScreenSaver: Make background black, and show Tux
If the video was black, the screensaver made it less dark. Set the
background color to black, not a very dark shade of gray.

My wife was complaining that the screensaver image was boring and
that she wanted a Pinguin back. So now Tux will appear on your TV.

Commit [81adbb] Authored by: Mike Looijmans 2015-01-25