TimerList: When ServiceName does not fit make it fit
So "Fox Sports Eredivisie Live HD" fits now. This also allow for timer
names to be longer then 2/3 rd of a line as it is now dynamic. But still
the timername is limited to be minimal 225 in width so you always see a
name

From Commit [a7f7ae] Authored by: littlesat 2014-12-22