streamserver.cpp: Set TCP send buffer to 188k and reducd receiving part
The streamserver won't be reading from the socket so it could close
the receiving end and free some resources. This resulted in aborting
the connection (socker notifier?), so instead just set the receiving
buffer to 1k.

It will typically write in 188k sized packets, so increase the TCP
send buffer to match that (a bug in the linux stack will actually
set it to double that size). Most send() requests should be able to
return immediately then, and it may reduce buffering problems in
clients.

From Commit [17bd02] Authored by: Mike Looijmans 2014-12-31