teletext subtitles: fix words sometimes being joined
together without proper whitespace.

Sometimes teletext subtitles are delivered with colour
attribute characters between the words, instead of spaces.
This is perfectly legal because on a hardware teletext
decoder, colour attributes also occupy a character cell.

Enigma2 appears to apply an optimisation where colour
attributes that change to the current colour (so actually
no change) are skipped. This is okay, but this makes the
colour attribute as whitespace to be dropped altogether,
without replacing them as whitespace.

Removed the optimisation altogether because it doesn't
make much sense anymore, it's probably a left-over from
(much) older code. The removal doesn't seem to break
anything, please report if it does.

From Commit [3fbbe1] Authored by: Erik Slagter 2015-01-11