epgcache.cpp: Remove the public lies from eventData.
All the members of eventData are being accessed directly in this file, so
don't lie and just name eventData a struct since it's all public anyway.

make CacheSize unsigned.

Write a note about the unaligned access (which proves that no one ever tried
to run this on an ARM which would barf at unaligned access attempts) but
don't address it because I intend to redo this descriptor stuff anyway.

From Commit [748bb7] Authored by: Mike Looijmans 2015-01-12