epgcache.cpp: Fix the unaligned uint32_t* access
creating a byte array and then casting that to an uint32_t* is very very bad,
in particular when you do that at a 10-byte offset.

Refactor the eventData struct to store the 10 bytes in one array, and the
list of CRC codes into another. This does not even increase the size of the
struct.

This solves the unaligned access, and also cleans up a few very ugly loops.
As a result, this code should now run correctly on e.g. ARM systems that cannot
do unaligned access, and it should run faster too on all systems.

It can even still read and write epg.dat files in the same format as before.

From Commit [3a4d42] Authored by: Mike Looijmans 2015-01-13