Make the size of the subtitles undependent from the skin size
e.g. in 1920x1080 skins the subtitles are very small in comparision with
1280x720 skins. Now the fontsize will be compensated so they look in size
the same in both type of skins. Also the border is taken into account.
So when changing to a 1920x1080 skin you do not need to alter the fontsize
for subtitles

From Commit [950b27] Authored by: Littlesat 2015-01-17