QuickSubtitleConfigMenu: Auto adjust screen size
When no fps are displayed resize the screen to the number of topics
available in the config list.

From Authored by: littlesat 2015-01-17

QuickSubtitleConfigMenu: Auto adjust screen size

When no fps are displayed resize the screen to the number of topics
available in the config list.

From Commit [a21f7f] Authored by: littlesat 2015-01-17