Do delete userbouquets when they are renamed
This solves the issue on http://forums.openpli.org/topic/36676-rename-bouquet/

When a bouquet was rename actually in the list it was removed and than pushed
back with the renamed name. Removing the bouquet actually add a .del to the
actual bouquet file on your box. Due to this this bouquet was not readed again
after an enigma2/box restart. By adding an option to removeService for not
adding the .del to the bouquet file and putting this default to true and
call it with False in ChannelSelection.py it can be avoided to add the .del
when we are renaming bouquets.

From Commit [2cd006] Authored by: Littlesat 2015-01-21