listboxservice: first try to solve anomaly when service list starts with number markers
When we use cable scan we also get a radio service, but the service numbers there do not
start with channel number 1. This leads to a side effect that you cannot correctly zap
through these radio channels via left/right. Only one direction works but it will skip
each time a channel. This patch at least solve the zap error. In fact the issue was that
when home was selected and the first service was a hidden service or hidden numbered marker
that this messes with the hole stuff as in fact this hidden service was still selected.
As we now search for the first not hidden service a hidden service cannot be selected
anymore. I tried this with empty list, list with only hidden markers, lists with only one
service and I could not find any further anomalies.

From Commit [6de524] Authored by: Littlesat 2015-01-22