SkinSelector: Only switch skin when you really do restart Enigma2
WHen you change the skin between e.g. a 720 and a 1080 after you
select the skin you get a message that the UI should be restarted
At this message you already get a png from the new skin. Now the skin
will be switched after you select yes on rebote so you will keep
the png there from the current selected skin. In addition we do
only switch the skin when you perform the reboot. So not store the
new skin and suddenly discover after a next reboot that the skin
was changed.

From Commit [7aac9b] Authored by: littlesat 2015-01-23