epgcache: Don't check on "empty" in cleanLoop
The chances of the eventDB being empty are very very slim, and the
iterator will run out of items quickly enough then. Remove the check
and move the lock to an inner scope, so that the lock is released
before setting the next timer.

From Commit [a3f56f] Authored by: Mike Looijmans 2015-01-24