epgcache.cpp: Remove Lock from public interface
Now that the internals are kept internal, there's no more need to
publish the implementation of the lock to the outside world.

Next up is to remove the evil "recursive" part from the lock so
that it can be replaced with a read/write lock that allows multiple
threads access to the cache if they're all just readers.

Also fix a minor const removal in a cast.

From Commit [9266ce] Authored by: Mike Looijmans Mike Looijmans 2015-01-24